Fixed product deletion. #331

+++ ecantina_project/api/views/product.py
@@ -5,6 +5,7 @@ from rest_framework import viewsets
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
 from rest_framework import filters
+from rest_framework import status
 from rest_framework.decorators import detail_route
 from api.pagination import TinyResultsSetPagination
 from api.permissions import BelongsToOrganizationOrReadOnly, BelongsToCustomerOrIsEmployeeUser, BelongsToCompanyPolicy
@@ -48,33 +49,70 @@ class ProductViewSet(viewsets.ModelViewSet):
     filter_backends = (filters.DjangoFilterBackend,)
     filter_class = ProductFilter
 
-    @detail_route(methods=['get'], permission_classes=[BelongsToCompanyPolicy])
-    def delete_with_additional_processing(self, request, pk=None):
+    def destroy(self, request, pk=None):
         """
-            Removes the product from the inventory permanetly and deletes the associated files.
-            Furthermore if any customer has this product in their cart, the cart gets automatically
-            updated to have totals done
+            Override the django-rest "delete" function to include custom
+            application logic.
         """
-        product = self.get_object() # Fetch the product we will be processing.
-        
-        # Fetch all the receipts which have this product in there.
-        receipts = Receipt.objects.filter(products__product_id=product.product_id)
+        # Step 1: Get the object
+        try:
+            product = Product.objects.get(product_id=int(pk))
+        except Product.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        # Step 2: Verify this product is not sold. If it's already sold then
+        #         we cannot delete this object.
+        if product.is_sold:
+            return Response({'status': 'failure', 'message': 'product cannot be deleted because it was already sold.'})
+    
+        # Step 3: Fetch all the (non-sold) receipts which have this product in
+        #         there and remove this product from the Receipt and then
+        #         recalculate the Receipt.
+        receipts = Receipt.objects.filter(products__product_id=product.product_id,has_finished=False,)
         for receipt in receipts:
-            receipt.products.remove(receipts)
-            # call_command('myadmincmd')
-            #TODO: Update
-            
-        # Delete the associated comic.
+            receipt.products.remove(product)
+            call_command('receipt_tally_up',str(receipt.receipt_id))
+        
+        # Step 4: Delete the associated objects and this object.
         try:
             Comic.objects.get(product=product).delete()
         except Exception as e:
             pass
-        
-        # Delete the phyiscal product.
         product.delete()
+                
+        # STEP 5: Update our Receipt (aka Cart).
+        call_command('receipt_tally_up',str(request.receipt.receipt_id))
 
-        # Return a success message.
-        return Response({'status': 'success', 'message': 'product was deleted from the database and the associated customer receipts have been updated.'})
+        # Step 6: Return a response.
+        return Response(status=status.HTTP_204_NO_CONTENT)
+#
+#    @detail_route(methods=['get'], permission_classes=[BelongsToCompanyPolicy])
+#    def delete_with_additional_processing(self, request, pk=None):
+#        """
+#            Removes the product from the inventory permanetly and deletes the associated files.
+#            Furthermore if any customer has this product in their cart, the cart gets automatically
+#            updated to have totals done
+#        """
+#        product = self.get_object() # Fetch the product we will be processing.
+#        
+#        # Fetch all the receipts which have this product in there.
+#        receipts = Receipt.objects.filter(products__product_id=product.product_id)
+#        for receipt in receipts:
+#            receipt.products.remove(receipts)
+#            # call_command('myadmincmd')
+#            #TODO: Update
+#            
+#        # Delete the associated comic.
+#        try:
+#            Comic.objects.get(product=product).delete()
+#        except Exception as e:
+#            pass
+#        
+#        # Delete the phyiscal product.
+#        product.delete()
+#
+#        # Return a success message.
+#        return Response({'status': 'success', 'message': 'product was deleted from the database and the associated customer receipts have been updated.'})
 
     @detail_route(methods=['get'], permission_classes=[BelongsToCustomerOrIsEmployeeUser])
     def apply_tax_and_discounts(self, request, pk=None):
